Fixed scoring.
The issue was,  cout<<setprecision(2)<<score would output in scientific
notation if the score was over 2 digits. At first I thought it was a
trash result, but then I relized it was giving me the right answer, just
not in normal notation. So I changed it to setprecision(4). I didn't
encounter this issue before because before I added in difficulty
specific scores, the highest score possible was 100, and I never got
that highest score before.